Add drakvuf-bundle customization
In some cases, users might want to add custom patches for the DRAKVUF
bundle. Add the ability to easily rebuild the deb by pointing a Docker
image at the repository and a directory with patches.

Current implementation allows only Xen and QEMU patching, however this
can be easily extended in the future.